split variation-actions-menu
