+++ packages/js/product-editor/src/components/variations-table/variation-actions/index.ts
@@ -1,3 +1,4 @@
-export * from './variation-actions-menu';
+export * from './quick-update-menu';
+export * from './single-update-menu';
 export * from './variation-actions-menu-item';
 export * from './types';
