+++ packages/js/product-editor/src/components/variations-table/variation-actions/quick-update-menu.tsx
@@ -1,0 +1,51 @@
+/**
+ * External dependencies
+ */
+import { Button, Dropdown } from '@wordpress/components';
+import { createElement } from '@wordpress/element';
+import { __ } from '@wordpress/i18n';
+import { chevronDown, chevronUp } from '@wordpress/icons';
+
+/**
+ * Internal dependencies
+ */
+import { VariationActionsMenuProps } from './types';
+import { QUICK_UPDATE } from './constants';
+import { VariationActions } from './variation-actions';
+
+export function QuickUpdateMenu( {
+	selection,
+	disabled,
+	onChange,
+	onDelete,
+}: VariationActionsMenuProps ) {
+	return (
+		<Dropdown
+			// @ts-expect-error missing prop in types.
+			popoverProps={ {
+				placement: 'bottom-end',
+			} }
+			renderToggle={ ( { isOpen, onToggle } ) => (
+				<Button
+					disabled={ disabled }
+					aria-expanded={ isOpen }
+					icon={ isOpen ? chevronUp : chevronDown }
+					variant="secondary"
+					onClick={ onToggle }
+					className="variations-actions-menu__toogle"
+				>
+					<span>{ __( 'Quick update', 'woocommerce' ) }</span>
+				</Button>
+			) }
+			renderContent={ ( { onClose }: { onClose: () => void } ) => (
+				<VariationActions
+					selection={ selection }
+					onClose={ onClose }
+					onChange={ onChange }
+					onDelete={ onDelete }
+					type={ QUICK_UPDATE }
+				/>
+			) }
+		/>
+	);
+}
